Don't set app id as application name
The application name is used to determine the data storage location and as we only have
one web process running in the background we can't use another location for every app.
Depending on the startup order the location was always different and apps couldn't access
their cookies / offline storages reliable.

Signed-off-by: Simon Busch <morphis@gravedo.de>