Restructure how we create our Qt window
If we create the window from QML it will not be destroyed at the right time. If we get the
close signal from the compositor we bring everything down but the window will be destroyed
after our application window class instance. Because of this the internal platform window
is recreated and we get a new surface for a short moment on the compositor side.

Signed-off-by: Simon Busch <morphis@gravedo.de>