Delete window platform resources directly
When we don't delete the platform resources first when destroying the window we will get a
surface registered again with the compositor which will bring a new card up for less than
a second but still visible. Destroying platform resources first will make sure the wayland
surface is destroyed and we can safely wait until we're ready to remove the window object
completely.

Signed-off-by: Simon Busch <morphis@gravedo.de>